Add legacy and worker auth action patterns
