+++ src/store/auth/index.js
@@ -1,25 +1,4 @@
-import LogRocket from 'logrocket'
-import { prefectAuth, prefectRefresh } from '@/middleware/prefectAuth'
 import jwt_decode from 'jwt-decode'
-
-import { OktaAuth } from '@okta/okta-auth-js'
-
-const {
-  VUE_APP_PUBLIC_CLIENT_ID,
-  VUE_APP_PUBLIC_ISSUER,
-  VUE_APP_ENVIRONMENT
-} = process.env
-
-const authClient = new OktaAuth({
-  clientId: VUE_APP_PUBLIC_CLIENT_ID,
-  issuer: VUE_APP_PUBLIC_ISSUER,
-  redirectUri: window.location.origin,
-  postLogoutRedirectUri: window.location.origin + '/logout',
-  scopes: ['openid', 'profile', 'email'],
-  testing: {
-    disableHttpsCheck: VUE_APP_ENVIRONMENT !== 'production'
-  }
-})
 
 const state = {
   authorizationToken: null,
@@ -266,184 +245,48 @@ const mutations = {
   }
 }
 
-const actions = {
-  async authenticate({ dispatch, commit, getters }) {
-    const urlParams = new URLSearchParams(window?.location?.search)
-    const invitationId = urlParams.get('invitation_id')
-    const source = urlParams.get('partner_source')
-    const error = urlParams.get('error')
-    // const errorDescription = urlParams.get('error_description') // We don't need to capture this at the moment
-    if (window.location?.pathname && !getters['redirectRoute']) {
-      dispatch(
-        'setRedirectRoute',
-        window.location.pathname + window.location.search
-      )
+let actions
+let worker
+
+try {
+  // // Initializes the shared service worker that handles token refresh
+  worker = new SharedWorker('../../workers/token.worker.js', {
+    type: 'module'
+  })
+
+  console.log(worker)
+
+  if (worker.port) {
+    worker.port.onmessage = e => {
+      console.log('Message received from worker', e)
     }
 
-    if (invitationId) {
-      sessionStorage.setItem('invitationId', invitationId)
+    worker.port.onmessageerror = e => {
+      console.log('Message received from worker', e)
     }
 
-    if (source) {
-      sessionStorage.setItem('partnerSource', source)
-    }
-
-    if (error) {
-      commit('error', error)
-      return
-    }
-
-    try {
-      const isAuthenticated = await authClient.isAuthenticated()
-      const isLoginRedirect = await authClient.isLoginRedirect()
-
-      const { tokens } = isLoginRedirect
-        ? await authClient.token.parseFromUrl()
-        : {
-            tokens: await authClient.tokenManager.getTokens()
-          }
-
-      if (tokens?.accessToken && tokens?.idToken) {
-        dispatch('commitTokens', tokens)
-        commit('isAuthenticated', true)
-      } else if (isAuthenticated) {
-        await dispatch('updateAuthentication')
-      } else {
-        commit('isAuthenticated', false)
-        await dispatch('login')
-      }
-    } catch (e) {
-      // eslint-disable-next-line
-      console.error(e)
-      commit('error', e)
-    }
-
-    return getters['isAuthenticated']
-  },
-  async authorize({ commit, getters, dispatch }) {
-    if (!getters['isAuthenticated']) return
-
-    commit('isAuthorizingUser', true)
-
-    const user = await authClient.getUser()
-    if (!user) return
-
-    commit('user', user)
-    dispatch('reportUserToLogRocket')
-
-    commit('user/setOktaUser', user, {
-      root: true
-    })
-
-    const prefectAuthorization = await prefectAuth(getters['idToken'])
-    if (prefectAuthorization) {
-      // Update authorization credentials if user is authorized
-      await dispatch('updateAuthorization', prefectAuthorization)
-    } else {
-      // Unset authorization if user is not authorized
-      commit('unsetAuthorizationToken')
-      commit('unsetAuthorizationTokenExpiry')
-      commit('unsetRefreshToken')
-      commit('unsetRefreshTokenExpiry')
-    }
-
-    commit('isAuthorizingUser', false)
-    return getters['authorizationToken']
-  },
-  async commitTokens({ commit }, tokens) {
-    const idToken = tokens.idToken
-    const accessToken = tokens.accessToken
-
-    authClient.tokenManager.setTokens(tokens)
-
-    commit('accessToken', accessToken.value)
-    commit('accessTokenExpiry', accessToken.expiresAt)
-
-    commit('idToken', idToken.value)
-    commit('idTokenExpiry', idToken.expiresAt)
-  },
-  async refreshAuthorization({ getters, commit, dispatch }) {
-    if (getters['isRefreshingAuthorization']) return
-
-    await dispatch('updateAuthentication')
-
-    commit('isRefreshingAuthorization', true)
-
-    const prefectAuthorization = await prefectRefresh(
-      getters['authorizationToken']
-    )
-
-    dispatch('updateAuthorization', prefectAuthorization)
-    commit('isRefreshingAuthorization', false)
-  },
-  async updateAuthorization({ commit }, authorization) {
-    commit('authorizationToken', authorization.access_token)
-    commit('refreshToken', authorization.refresh_token)
-    commit(
-      'authorizationTokenExpiry',
-      new Date(authorization.expires_at).getTime()
-    )
-    commit(
-      'refreshTokenExpiry',
-      jwt_decode(authorization.refresh_token).exp * 1000
-    )
-  },
-  async updateAuthentication({ dispatch, commit, getters }) {
-    if (getters['isRefreshingAuthentication']) return
-
-    commit('isRefreshingAuthentication', true)
-
-    try {
-      // This should manually update authentication
-      const idToken = await authClient.tokenManager.get('idToken')
-      const accessToken = await authClient.tokenManager.get('accessToken')
-
-      if (!idToken || !accessToken) {
-        commit('isAuthenticated', false)
-        await dispatch('login')
-      } else {
-        commit('isAuthenticated', true)
-        dispatch('commitTokens', {
-          idToken: idToken,
-          accessToken: accessToken
-        })
-      }
-    } catch (e) {
-      // eslint-disable-next-line
-      console.error(e)
-      commit('error', e)
-    }
-
-    commit('isRefreshingAuthentication', false)
-  },
-  async login({ commit }) {
-    commit('isLoggingInUser', true)
-    authClient.token.getWithRedirect({
-      responseType: ['token', 'id_token']
-    })
-    commit('isLoggingInUser', false)
-  },
-  async logout({ commit }) {
-    commit('isAuthenticated', false)
-    commit('unsetIdToken')
-    commit('unsetAccessToken')
-    commit('unsetRedirectRoute')
-    await authClient.signOut()
-  },
-  setRedirectRoute({ commit }, redirectRoute) {
-    commit('redirectRoute', redirectRoute)
-  },
-  removeRedirectRoute({ commit }) {
-    commit('unsetRedirectRoute')
-  },
-  reportUserToLogRocket({ state }) {
-    if (!state.user) return
-
-    LogRocket.identify(state.user.sub, {
-      name: state.user.name || '',
-      email: state.user.email || ''
-    })
+    let start = worker.port.start()
+    console.log(start)
   }
+
+  worker.onmessage = e => {
+    console.log('Message received from worker', e)
+  }
+
+  worker.onmessageerror = e => {
+    console.log('Message received from worker', e)
+  }
+
+  actions = require('./worker-actions.js').default.actions
+
+  throw new Error('testing dynamic import/export')
+} catch (e) {
+  console.log(
+    'Unable to install Shared Service worker, falling back to legacy auth pattern.',
+    e
+  )
+
+  actions = require('./legacy-actions.js').default.actions
 }
 
 export default {
@@ -453,3 +296,5 @@ export default {
   actions,
   namespaced: true
 }
+
+export { worker }
