+++ src/store/auth/worker-actions.js
@@ -1,0 +1,5 @@
+const actions = {}
+
+export default {
+  actions
+}
