+++ src/vue-apollo.js
@@ -20,17 +20,17 @@ Vue.use(VueApollo)
 // Name of the localStorage item
 const AUTH_TOKEN = 'authorization_token'
 
-function isExpired(expiry) {
-  return new Date().getTime() > expiry
-}
+// function isExpired(expiry) {
+//   return new Date().getTime() > expiry
+// }
 
-function notExpired(expiry) {
-  return !isExpired(expiry)
-}
+// function notExpired(expiry) {
+//   return !isExpired(expiry)
+// }
 
-function aboutToExpire(expiry) {
-  return notExpired(expiry) && new Date().getTime() + 5000 >= expiry
-}
+// function aboutToExpire(expiry) {
+//   return notExpired(expiry) && new Date().getTime() + 5000 >= expiry
+// }
 
 let errors = 0,
   apiErrors = 0
@@ -181,33 +181,35 @@ const authMiddleware = setContext(async (_, { headers }) => {
     }
   }
 
-  const authRefreshRequired =
-    store.getters['auth/authorizationToken'] &&
-    aboutToExpire(store.getters['auth/authorizationTokenExpiry'])
+  // const authRefreshRequired =
+  //   store.getters['auth/authorizationToken'] &&
+  //   aboutToExpire(store.getters['auth/authorizationTokenExpiry'])
 
-  const validRefreshToken =
-    store.getters['auth/refreshToken'] &&
-    notExpired(store.getters['auth/refreshTokenExpiry'])
+  // const validRefreshToken =
+  //   store.getters['auth/refreshToken'] &&
+  //   notExpired(store.getters['auth/refreshTokenExpiry'])
 
-  const isAuthenticatedUser =
-    store.getters['auth/idToken'] &&
-    notExpired(store.getters['auth/idTokenExpiry'])
+  // const isAuthenticatedUser =
+  //   store.getters['auth/idToken'] &&
+  //   notExpired(store.getters['auth/idTokenExpiry'])
 
-  const middleOfRefresh =
-    store.getters['auth/isRefreshingAuthorization'] ||
-    store.getters['auth/isAuthorizingUser'] ||
-    store.getters['auth/isLoggingInUser']
+  // const middleOfRefresh =
+  //   store.getters['auth/isRefreshingAuthorization'] ||
+  //   store.getters['auth/isAuthorizingUser'] ||
+  //   store.getters['auth/isLoggingInUser']
 
-  if (authRefreshRequired && !middleOfRefresh) {
-    defaultApolloClient.cache.reset()
-    if (validRefreshToken) {
-      await store.dispatch('auth/refreshAuthorization')
-    } else if (isAuthenticatedUser) {
-      await store.dispatch('auth/authorize')
-    } else {
-      await store.dispatch('auth/login')
-    }
-  }
+  // if (authRefreshRequired && !middleOfRefresh) {
+  //   // console.log(authRefreshRequired, _.operationName)
+  //   defaultApolloClient.cache.reset()
+  //   if (validRefreshToken) {
+  //     // console.log(validRefreshToken, _.operationName)
+  //     await store.dispatch('auth/refreshAuthorization')
+  //   } else if (isAuthenticatedUser) {
+  //     await store.dispatch('auth/authorize')
+  //   } else {
+  //     await store.dispatch('auth/login')
+  //   }
+  // }
 
   const bearer = `Bearer ${store.getters['auth/authorizationToken']}`
   return {
@@ -225,6 +227,13 @@ const link = ApolloLink.from([
   backendMiddleware,
   errorAfterware,
   batchLink
+])
+
+const authLink = ApolloLink.from([
+  authMiddleware,
+  headerMiddleware,
+  backendMiddleware,
+  errorAfterware
 ])
 
 export const cache = new InMemoryCache({
@@ -308,6 +317,14 @@ export { defaultApolloClient }
 export const fallbackApolloProvider = createApolloProvider()
 fallbackApolloClient = fallbackApolloProvider.defaultClient
 export { fallbackApolloClient }
+
+export const authApolloProvider = createApolloProvider({
+  ...defaultOptions,
+  link: authLink,
+  defaultHttpLink: true
+})
+const authApolloClient = authApolloProvider.defaultClient
+export { authApolloClient }
 
 export async function apolloOnLogin(apolloClient) {
   if (apolloClient.wsClient) restartWebsockets(apolloClient.wsClient)
